1、支持多张自制的暂估应付单生成一张发票 2、发票取消传应付时，取消回冲应付单错误，没有将对应的回冲应付单取消，而是将另外一张回冲应付单取消了。 3、.不同类型，同一供应商的暂估应付单参照生成发票，合并
git-svn-id: http://59.39.61.138:8088/svn/junjie@1043 d7022953-1ca3-a944-a70e-6a56d19ef69b

+++ junjie/develop/proj/newmodules/pu/src/private/nc/bs/pi/InvoiceImpl.java
@@ -7512,7 +7512,7 @@ public void adjustForFeeZGYF(InvoiceVO[] voaInv) throws BusinessException {
 				if (numtem != null
 						&& new UFDouble(numtem.toString()).toDouble() < 0) {
 					throw new BusinessException("第" + itemvo[j].getCrowno()
-							+ "行的累计开单数量已超出可开单数量:"+new UFDouble(numtem.toString()).toDouble()+"，不能保存，请检查！");
+							+ "行的累计开单数量已超出可开单数量:"+new UFDouble(numtem.toString())+"，不能保存，请检查！");
 				}
 			}
